NXP-23065: move KeyValueService to its own bundle
