+++ nuxeo-core/nuxeo-core-kv/src/test/java/org/nuxeo/ecm/core/kv/AbstractKeyValueStoreTest.java
@@ -16,7 +16,7 @@
  * Contributors:
  *     Florent Guillaume
  */
-package org.nuxeo.ecm.core.storage.kv;
+package org.nuxeo.ecm.core.kv;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.Assert.assertArrayEquals;
