+++ nuxeo-core/nuxeo-core-kv/src/test/java/org/nuxeo/ecm/core/kv/TestMemKeyValueStore.java
@@ -16,7 +16,7 @@
  * Contributors:
  *     Florent Guillaume
  */
-package org.nuxeo.ecm.core.storage.kv;
+package org.nuxeo.ecm.core.kv;
 
 /**
  * @since 9.1
