+++ nuxeo-core/nuxeo-core-storage/pom.xml
@@ -74,10 +74,6 @@
       <artifactId>guava</artifactId>
     </dependency>
     <dependency>
-      <groupId>net.jodah</groupId>
-      <artifactId>expiringmap</artifactId>
-    </dependency>
-    <dependency>
       <groupId>net.htmlparser.jericho</groupId>
       <artifactId>jericho-html</artifactId>
     </dependency>
