+++ nuxeo-core/nuxeo-core-storage/src/main/resources/META-INF/MANIFEST.MF
@@ -1,5 +1,4 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.nuxeo.ecm.core.storage
-Nuxeo-Component: OSGI-INF/lockmanager-service-contrib.xml,
- OSGI-INF/keyvalue-service.xml
+Nuxeo-Component: OSGI-INF/lockmanager-service-contrib.xml
